% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-math.R
\name{expit}
\alias{expit}
\title{Logistic function}
\usage{
expit(x)
}
\arguments{
\item{x}{Numeric vector on the log-odds scale.}
}
\value{
Numeric vector of probabilities between 0 and 1.
}
\description{
Computes the inverse-logit transformation.
}
\examples{
expit(0)
expit(c(-2, 0, 2))

}
