% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-math.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{Numeric vector of probabilities between 0 and 1.}
}
\value{
Numeric vector on the log-odds scale.
}
\description{
Computes the logit transformation of probabilities.
}
\examples{
logit(0.5)
logit(c(0.1, 0.5, 0.9))

}
